Fix GPUSet running on CPU.
* Move GPUSet into common.h
* Split AllVisible definition.
