+++ src/common/common.cc
@@ -1,9 +1,11 @@
 /*!
- * Copyright 2015 by Contributors
+ * Copyright 2015-2018 by Contributors
  * \file common.cc
  * \brief Enable all kinds of global variables in common.
  */
 #include <dmlc/thread_local.h>
+
+#include "common.h"
 #include "./random.h"
 
 namespace xgboost {
@@ -20,4 +22,20 @@ GlobalRandomEngine& GlobalRandom() {
   return RandomThreadLocalStore::Get()->engine;
 }
 }  // namespace common
+
+#ifndef XGBOOST_USE_CUDA
+GPUSet GPUSet::All(int ndevices, int num_rows) {
+  return Empty();
+}
+
+GPUSet GPUSet::AllVisible() {
+  return Empty();
+}
+
+int GPUSet::GetDeviceIdx(int gpu_id) {
+  LOG(FATAL) << "Not part of device code";
+  return 0;
+}
+#endif
+
 }  // namespace xgboost
